Directorystructure refactored
