+++ UndoRedoFramework/include/UndoRedoFramework/CompositeCommandImpl.h
